+++ functions/model/payment/tark/generateRazorpayOrder.js
@@ -9,7 +9,7 @@ const { standardSubscription } = require("../../application/lib");
 const { getSubscriptions } = require("../../subscriptions/lib");
 const { setRazorpayOrderDetails, generateBase62Constant, generatePaymentId} = require("../lib");
 const { currentTime, currentDate } = require("../../application/lib");
-const { keyId, keySecret, orderKey} = require("../../../paymentKeys");
+const { keyId, keySecret, orderKey} = require("../../application/paymentKeys");
 exports.generateRazorpayOrder = function(request, response) {
         const Uid = request.body.data.Uid;
         const SubscriptionId = request.body.data.SubscriptionId;
